individual project view renders and delete project function works
