+++ public/css/style.css
@@ -63,6 +63,12 @@ a {
   text-decoration: underline;
 }
 
+.link-to-project{
+  width: 100%;
+  height: 100%;
+  text-decoration: none;
+}
+
 header {
   display: flex;
   align-items: center;
@@ -78,6 +84,10 @@ header a {
   color: var(--contrast);
   font-size: 1.25rem;
   text-decoration: none;
+}
+
+#projectIDblind{
+  display: none;
 }
 
 #logo {
