+++ routes/dashboard.js
@@ -7,7 +7,8 @@ const {ensureAuth, ensureGuest} = require('../middleware/auth')
 
 router.get('/', ensureAuth, dashController.getDash)
 router.get('/addProject', dashController.addProject)
+router.get('/project', projectController.getProject)
 // router.post('/postProject', dashController.postProject)
-// router.delete('/delete', dashController.deleteProject)
+router.delete('/delete', projectController.deleteProject)
 
 module.exports = router
