+++ routes/projectDetail.js
@@ -2,6 +2,7 @@ const express = require('express')
 const router = express.Router()
 const projController = require('../controllers/project')
 
+// router.get('/project', projController.getProject)
 router.post('/markComplete', projController.markComplete)
 router.post('/markIncomplete', projController.markIncomplete)
 router.delete('/projectDelete', projController.deleteProject)
